requirements.txt
transformers
torch
pillow
reportlab
huggingface_hub
accelerate
